which refactor p1
