+++ xonsh/aliases.py
@@ -22,7 +22,7 @@ from xonsh.timings import timeit_alias
 from xonsh.tools import (XonshError, argvquote, escape_windows_cmd_string,
                          to_bool)
 from xonsh.xontribs import xontribs_main
-from xonsh.xoreutils import _which
+from xonsh.xoreutils.which import which_main
 
 import xonsh.completers._aliases as xca
 
@@ -362,112 +362,6 @@ class AWitchAWitch(argparse.Action):
         parser.exit()
 
 
-def which(args, stdin=None, stdout=None, stderr=None):
-    """
-    Checks if each arguments is a xonsh aliases, then if it's an executable,
-    then finally return an error code equal to the number of misses.
-    If '-a' flag is passed, run both to return both `xonsh` match and
-    `which` match.
-    """
-    desc = "Parses arguments to which wrapper"
-    parser = argparse.ArgumentParser('which', description=desc)
-    parser.add_argument('args', type=str, nargs='+',
-                        help='The executables or aliases to search for')
-    parser.add_argument('-a', '--all', action='store_true', dest='all',
-                        help='Show all matches in $PATH and xonsh.aliases')
-    parser.add_argument('-s', '--skip-alias', action='store_true',
-                        help='Do not search in xonsh.aliases', dest='skip')
-    parser.add_argument('-V', '--version', action='version',
-                        version='{}'.format(_which.__version__),
-                        help='Display the version of the python which module '
-                        'used by xonsh')
-    parser.add_argument('-v', '--verbose', action='store_true', dest='verbose',
-                        help='Print out how matches were located and show '
-                        'near misses on stderr')
-    parser.add_argument('-p', '--plain', action='store_true', dest='plain',
-                        help='Do not display alias expansions or location of '
-                             'where binaries are found. This is the '
-                             'default behavior, but the option can be used to '
-                             'override the --verbose option')
-    parser.add_argument('--very-small-rocks', action=AWitchAWitch)
-    if ON_WINDOWS:
-        parser.add_argument('-e', '--exts', nargs='*', type=str,
-                            help='Specify a list of extensions to use instead '
-                            'of the standard list for this system. This can '
-                            'effectively be used as an optimization to, for '
-                            'example, avoid stat\'s of "foo.vbs" when '
-                            'searching for "foo" and you know it is not a '
-                            'VisualBasic script but ".vbs" is on PATHEXT. '
-                            'This option is only supported on Windows',
-                            dest='exts')
-    if len(args) == 0:
-        parser.print_usage(file=stderr)
-        return -1
-    pargs = parser.parse_args(args)
-
-    if pargs.all:
-        pargs.verbose = True
-
-    if ON_WINDOWS:
-        if pargs.exts:
-            exts = pargs.exts
-        else:
-            exts = builtins.__xonsh_env__['PATHEXT']
-    else:
-        exts = None
-
-    failures = []
-    for arg in pargs.args:
-        nmatches = 0
-        # skip alias check if user asks to skip
-        if (arg in builtins.aliases and not pargs.skip):
-            if pargs.plain or not pargs.verbose:
-                if not callable(builtins.aliases[arg]):
-                    print(' '.join(builtins.aliases[arg]), file=stdout)
-                else:
-                    print(arg, file=stdout)
-            else:
-                print("aliases['{}'] = {}".format(arg, builtins.aliases[arg]), file=stdout)
-                if callable(builtins.aliases[arg]):
-                    builtins.__xonsh_superhelp__(builtins.aliases[arg])
-            nmatches += 1
-            if not pargs.all:
-                continue
-        # which.whichgen gives the nicest 'verbose' output if PATH is taken
-        # from os.environ so we temporarily override it with
-        # __xosnh_env__['PATH']
-        original_os_path = os.environ['PATH']
-        os.environ['PATH'] = builtins.__xonsh_env__.detype()['PATH']
-        matches = _which.whichgen(arg, exts=exts, verbose=pargs.verbose)
-        for abs_name, from_where in matches:
-            if ON_WINDOWS:
-                # Use list dir to get correct case for the filename
-                # i.e. windows is case insensitive but case preserving
-                p, f = os.path.split(abs_name)
-                f = next(s.name for s in scandir(p) if s.name.lower() == f.lower())
-                abs_name = os.path.join(p, f)
-                if builtins.__xonsh_env__.get('FORCE_POSIX_PATHS', False):
-                    abs_name.replace(os.sep, os.altsep)
-            if pargs.plain or not pargs.verbose:
-                print(abs_name, file=stdout)
-            else:
-                print('{} ({})'.format(abs_name, from_where), file=stdout)
-            nmatches += 1
-            if not pargs.all:
-                break
-        os.environ['PATH'] = original_os_path
-        if not nmatches:
-            failures.append(arg)
-    if len(failures) == 0:
-        return 0
-    else:
-        print('{} not in $PATH'.format(', '.join(failures)), file=stderr, end='')
-        if not pargs.skip:
-            print(' or xonsh.builtins.aliases', file=stderr, end='')
-        print('', end='\n')
-        return len(failures)
-
-
 def xonfig(args, stdin=None):
     """Runs the xonsh configuration utility."""
     from xonsh.xonfig import xonfig_main  # lazy import
@@ -534,7 +428,7 @@ def make_default_aliases():
         'scp-resume': ['rsync', '--partial', '-h', '--progress', '--rsh=ssh'],
         'showcmd': showcmd,
         'ipynb': ['jupyter', 'notebook', '--no-browser'],
-        'which': which,
+        'which': which_main,
         'xontrib': xontribs_main,
         'completer': xca.completer_alias
     }
